code_all/7_code_clothing/requirements_clothing.txt
# Clothing Analysis Requirements
# Add these to your main requirements.txt or install separately

# Vision-Language Models
transformers>=4.30.0
accelerate>=0.20.0
protobuf>=3.20.0

# Image Processing
pillow>=9.0.0

# PyTorch (already in main requirements.txt for Blackwell support)
# torch==2.7.0
# torchvision==0.22.0
# torchaudio==2.7.0

# Additional dependencies for InstructBLIP
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Optional: For faster inference
# torch-audio>=2.7.0  # Already included in main requirements
